requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
alembic
passlib[bcrypt]
python-multipart
pdfplumber
openai
requests
schedule
chromadb
tiktoken
python-jose[cryptography]